Move extension function into it's own file.
